requirements.txt
click==8.0.3
PyYAML==6.0
pytest==7.0.1
sphinx==4.4.0
